fix: rename Internal namespace
