+++ packages/shared/src/index.ts
@@ -6,4 +6,3 @@ export * from './utils/env';
 export * from './utils/testUtils';
 
 export * as internalApi from './internalApi/api';
-export * from './internalApi/types';
